+++ ccc_client/app_repo/cli/list_tools.py
@@ -1,0 +1,8 @@
+# v2/_catalog
+import argparse
+
+def run(args):
+    r = runner.list_tools()
+
+parser = argparse.ArgumentParser()
+parser.set_defaults(runner=run)
