+++ ccc_client/elastic_search/__init__.py
@@ -1,0 +1,1 @@
+from . import cli
